refactor: unify roadview game view and split flow services
