+++ src/router/devRoutes.js
@@ -4,7 +4,7 @@ import kakaoMapTest from '@/dummy/KakaoMapTest.vue'
 import EncryptTest from 'src/dummy/EncryptTest.vue'
 // import ComponentTest from 'src/dummy/ComponentTest.vue'
 import TempLoginView from '@/features/auth/views/TempLoginView.vue'
-import SoloGameView from '@/features/game/multiplayer/roadview/views/SoloGameView.vue'
+import SoloGameView from '@/features/game/multiplayer/roadview/views/BaseGameView.vue'
 import RoomView from '@/features/game/multiplayer/room/views/RoomView.vue'
 import DevTestPage from '@/dummy/DevTestPage.vue'
 import ResultOverlay from '@/features/game/single/roadview/components/Result/ResultOverlay.vue'
